+++ src/oihana/core/strings/between.php
@@ -17,7 +17,7 @@ namespace oihana\core\strings;
  *
  * @package oihana\core\strings
  * @author  Marc Alcaraz
- * @since   1.0.0
+ * @since   1.0.6
  *
  * @example
  * ```php